Added some compile time asserts.

+++ LCG.h
@@ -1,4 +1,18 @@
-#include "compression-api.h"
+// ms-compress: implements Microsoft compression algorithms
+// Copyright (C) 2012  Jeffrey Bush  jeff@coderforlife.com
+//
+// This library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 // Tabulation Hashing
 // Will hash 3 bytes at a time to a number from 0 to MAX_HASH-1 (inclusive)
@@ -23,9 +37,13 @@
 //			}
 //			table[j] = (hash -= 0x789ABCDFu) & (MAX_HASH - 1);
 
-template<uint32_t SEED, uint32_t A, uint32_t C, uint32_t M, uint32_t S, uint32_t MaxHash>
+#include "compression-api.h"
+
+template<uint32_t SEED, uint32_t A, uint32_t C, uint32_t M, uint32_t S, uint16_t MaxHash>
 class LCG
 {
+	//TODO: CASSERT(IS_POW2(MaxHash));
+
 private:
 	uint16_t hashes[3][0x100];
 
@@ -59,5 +77,5 @@ public:
 	inline static uint_fast16_t Hash(const const_bytes x) { return lcg.hashes[0][x[0]] ^ lcg.hashes[1][x[1]] ^ lcg.hashes[2][x[2]]; }
 };
 
-template<uint32_t SEED, uint32_t A, uint32_t C, uint32_t M, uint32_t S, uint32_t MaxHash>
+template<uint32_t SEED, uint32_t A, uint32_t C, uint32_t M, uint32_t S, uint16_t MaxHash>
 LCG<SEED, A, C, M, S, MaxHash> LCG<SEED, A, C, M, S, MaxHash>::lcg;
